requirements.txt
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.1
black==21.11b1
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
coverage==6.1.2
cryptography==35.0.0
distlib==0.3.3
Django==3.2.9
djangorestframework==3.12.4
filelock==3.4.0
gunicorn==20.1.0
identify==2.4.0
idna==3.3
iniconfig==1.1.1
model-bakery==1.3.3
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
pip==21.3.1
platformdirs==2.4.0
pluggy==1.0.0
pre-commit==2.15.0
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
PySocks==1.7.1
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.4.0
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
setuptools==59.2.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.0
urllib3==1.26.7
virtualenv==20.10.0
wheel==0.37.0
